requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
Flask-SQLAlchemy==3.0.5
SQLAlchemy==2.0.19
psycopg2-binary>=2.9.9
PyJWT==2.8.0
